添加搜索功能,so正常了

+++ reader/src/main/java/cn/archko/pdf/widgets/APDFView.kt
@@ -45,7 +45,7 @@ class APDFView(
         val paint = Paint()
         paint.color = Color.BLUE
         paint.isAntiAlias = true
-        paint.textSize = Utils.sp2px(30f).toFloat()
+        paint.textSize = Utils.sp2px(30f)
         paint.textAlign = Paint.Align.CENTER
         return paint
     }
